real-world-projects/5.fastapi-with-ai/src/core/token.py
from datetime import datetime, timedelta, timezone

from fastapi import HTTPException  # pyright: ignore[reportMissingImports]
import jwt  # pyright: ignore[reportMissingImports]
from jwt.exceptions import InvalidTokenError, ExpiredSignatureError  # pyright: ignore[reportMissingImports]

from src.core.config import settings
from src.schemas.auth_schema import TokenData



SECRET_KEY = settings.SECRET_KEY    
ALGORITHM = settings.ALGORITHM
ACCESS_TOKEN_EXPIRE_MINUTES = settings.ACCESS_TOKEN_EXPIRE_MINUTES


def create_access_token(data: dict, expires_delta: timedelta | None = None):
    to_encode = data.copy()
    if expires_delta:
        expire = datetime.now(timezone.utc) + expires_delta
    else:
        expire = datetime.now(timezone.utc) + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
    return encoded_jwt


def verify_access_token(token: str, credentials_exception):
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        email = payload.get("sub")
        if email is None:
            raise credentials_exception
        return TokenData(email=email)
    except ExpiredSignatureError:
        raise HTTPException(status_code=401, detail="Token has expired")
    except InvalidTokenError:
        raise credentials_exception
    